storage c4_Storage ctor second parameter is of type int with values 0 or 1
Fix PVS-Studio issue "V601 The 'true' value is implicitly cast to the
integer type. Inspect the second argument. feedstorage.cpp 103".

+++ src/storage/feedstorage.cpp
@@ -100,7 +100,7 @@ FeedStorage::FeedStorage(const QString &url, Storage *main)
     QString t = url2;
     QString t2 = url2;
     QString filePath = main->archivePath() + QLatin1Char('/') + t.replace(QLatin1Char('/'), QLatin1Char('_')).replace(QLatin1Char(':'), u'_');
-    d->storage = new c4_Storage(QString(filePath + QLatin1StringView(".mk4")).toLocal8Bit().constData(), true);
+    d->storage = new c4_Storage(QString(filePath + QLatin1StringView(".mk4")).toLocal8Bit().constData(), static_cast<int>(true));
 
     d->archiveView = d->storage->GetAs(
         "articles[guid:S,title:S,hash:I,guidIsHash:I,guidIsPermaLink:I,description:S,link:S,comments:I,commentsLink:S,status:I,pubDate:I,tags[tag:S],"
